Added a pre-compile script for machine support code.
Each machine has a .js, .lua or .py file that supports the interop with the engine, I've added a pre-build shell script to generate a header file for each now so that they can be worked on with intellisense. Also msvcc is now being used to compile on windows instead of gcc.